Crate tracing_web

A tracing compatible subscriber layer for web platforms.

Example usage

use tracing_web::{MakeWebConsoleWriter, performance_layer};
use tracing_subscriber::fmt::format::Pretty;
use tracing_subscriber::fmt::time::UtcTime;
use tracing_subscriber::prelude::*;

let fmt_layer = tracing_subscriber::fmt::layer()
    .with_ansi(false) // Only partially supported across browsers
    .without_time()   // std::time is not available in browsers
    .with_writer(MakeWebConsoleWriter::new()); // write events to the console
let perf_layer = performance_layer().with_details_from_fields(Pretty::default());

tracing_subscriber::registry()
    .with(fmt_layer)
    .with(perf_layer)
    .init();

Structs

ConsoleWriter

Concrete std::io::Write implementation returned by MakeConsoleWriter and MakeWebConsoleWriter.

FormatSpanFromFields

An adaptor for Formatters from tracing_subscriber::fmt::format as a FormatSpan.

MakeConsoleWriter

Discouraged A MakeWriter emitting the written text to the console.

MakeWebConsoleWriter

A MakeWriter emitting the written text to the console.

PerformanceEventsLayer

A Layer that emits span enter, exit and events as performance marks.

Traits

FormatSpan

Determine what additional information will be attached to the performance events.

Functions

performance_layer

Construct a new layer recording performance events.