tracing_web

Struct MakeWebConsoleWriter

pub struct MakeWebConsoleWriter { /* private fields */ }

A MakeWriter emitting the written text to the console.

The used log method is sensitive to the level the event is emitted with.

Level	Method

TRACE	console.debug

DEBUG	console.debug

INFO	console.info

WARN	console.warn

ERROR	console.error

other	console.log

Implementations

impl MakeWebConsoleWriter

pub fn new() -> Self

Create a default console writer, i.e. no level annotation is shown when logging a message.

pub fn with_pretty_level(self) -> Self

Enables an additional label for the log level to be shown.

It is recommended that you also use Layer::with_level(false) if you use this option, to avoid the event level being shown twice.